An overlay operating system which allows users to use a  new custom interface instead of a boring windows interface
This project basically allows users to change the interface in which
they use on Windows PCs, it works like an operating system however it is
just an overlay at the backend your windows system is doing the real
work. I  made this mainly to practice a graphical feel of creating an
operating system.